Work with yarn workspaces
Having all the stuff installed for each element was getting a bit
overwhelming in terms of disk space. I'm switching over to using
Yarn workspaces to reduce disk usage.

This has created a new problem for the tests though. Since the workspace
node_modules/ folder has a bunch of the packages now, the testing
tool's attempts to look them up from the node_modules folder is failing.
wct looks like it's trying to resolve packages, but it's looking in
the wrong place for some reason.