Test with Karma and Mocha
Using wct for testing wasn't working with yarn workspaces, so i've
switched over to using a combination of Karma and Mocha. I've converted
the tests for web3-element, and will do web3-accounte-element shortly.

While I was troubleshooting, I learned about named slots in the shadowdom.
This means we can have a slot named "noweb3", and an element in the light
dom that attaches to the "noweb3" slot. So rather than have a separate
noweb3 tag (which could potentially get out of sync with the web3 tag)
the noweb3 option will be a named slot within the web3 tag.