Add fee element
The fee element queries the SRA for the maker and taker fee, and
gives the user the ability to adust the distribution. It starts out
with the initial distribution provided by the SRA, and the element
can be disabled to prevent the user from changing the fees while
still displaying the distribution.