Network support
Add a mechanism for specifying whicn networks are supported and which are not.

The DOM can specify tags with 'slot="net"' and the inner text of the element
indicating a supported network. If no such tags are specified, all networks
are assumed to be supported. If any net tags are specified, the selected network
must be one of the specified networks.

It also adds a netunsupported slot, which displays if the user is on an unsupported
network, allowing the dapp to specify supported networks.