Debounce loading
Since web3 injection takes a moment, we were seeing the "noweb3" slot
for a brief moment before the web3 injection came in and replaced it
with the standard slot. This adds a "loaded" flag that gets set when
web3 injection is complete, or after 250 ms if no web3 is found, to
prevent the brief flash of noweb3 content.