Fix token issues
There were a couple of issues relating to token selection / symbols.

For the token selector, it was only displaying greenlist items, but
selecting based on indexes in the token list. So if you selected WETH
(the 4th item in the greenlist) for example, it would provide data for
AFA (the 4th item in the complete list). This was fixed in the token
selector.

For sra-signed-order, it was failing to display DAI because the DAI contract
implements the Symbol view differently than most ERC20 contracts.
Now sra-signed-order will only hit the symbol view if the token address
is not found in our managed token list.